staticanalysis/bot: Move Phabricator API to cli_common
